requirements.txt
Flask==3.1.0
faiss-cpu==1.11.0
datasets==3.6.0
numpy==2.2.5
openai==1.76.0
tqdm==4.67.1
